Remove CMake option URL_TEST_COROUTINE
For Visual Studio 2019 16.7 and earlier this option added compiler
switch /await. As of VS 2019 16.8 this switch is not required anymore to
use coroutines, it is sufficient to specify C++20 standard
(/std:c++latest). Also this defines __cpp_impl_coroutine macro.

This commit removes this option, and checking for old __cpp_coroutines
macro.

More info:
https://devblogs.microsoft.com/cppblog/c-coroutines-in-visual-studio-2019-version-16-8/