Remove file_path_format::detect
Use `file_path_format::native` as default in the `url_from_file_path`
function.

The path format detection algorithm was simple: if the first path
character is `/`, then it is POSIX format, otherwise - Windows. But in
some cases it is ambiguous, e.g. path `//h\sn\p` is recognized as POSIX
path, but it can also be Windows UNC path, because `/` is allowed as
directory separator in Windows paths.