Fix a test in IdnaTestV2.json from web-platform-tests
The test in question is:

    {
      "comment": "V2 (ignored)",
      "input": "xn--xn--a--gua.pt",
      "output": "xn--xn--a--gua.pt"
    }

ToASCII with this "input" value must return a failure value due to the
new V4 validity criteria in Unicode IDNA 15.1. So the "output" value
must be null.

The fixed test is in the IdnaTestV2-fixes.json file and is used instead
of the invalid test.

See: https://www.unicode.org/reports/tr46/tr46-31.html#Validity_Criteria