Move code that accesses public_suffix_list::root_ to .cpp file
The variable `root_` is of type `label_item` struct, containing a member
named `children`. The type of `children` differs depending on whether
the code is compiled with C++17 or C++20. It is possible to encounter a
situation where the Upa URL library is compiled using a C++17 compiler,
but is then integrated into a program compiled with C++20. In such
cases, it is important that all code accessing `root_` is compiled using
the same C++ standard, prompting the implementation of this fix.